test: use py3{py-minor-version} format rather than py3.{py-minor-version} for testenv name
This convention is more "tox-friendly", since an environment with 'py$MAJOR$MINOR'
in its name will use the Python interpreter at version $MAJOR.MINOR as base python.

In particular, in this way, we could avoid the repetition of different testenv
configurations which only differ by the basepython variable.

This change required a workaround in the GitHub Actions workflow configuration,
since the Python version there has a dot that separates major and minor versions.