feat: add TypesIndex class
This commit adds the parsing utility class "TypesIndex".

TypesIndex is an index for PDDL types and PDDL names/variables. It is used to index PDDL names and variables by their types. OrderedDict is used to preserve the order of the types and the names, e.g. for predicate variables. Other types of validations are performed to ensure that the types index is consistent.

In this commit, it is used only to parse typed lists of names (not variables).