fix: improve Problem initialize regarding requirements and domain_name vs domain
Problem objects can be instantiated in two ways: either with a domain object, or without it.

If with a domain object, then both the attributes/args domain_name and requirements must be validated against the provided domain.
Otherwise, the provided arguments are used to set the attributes.

The domain setter will reset both requirements and domain_name attributes.
The property getter of requirements will return empty set even when it was not specified (therefore, interpreting absence of requirements section as empty set of requirements).