fix: move keyword validation inside class contructors
This commit adds keyword validation for 'name' types. In particuar, instead of using the raw 'name' constructor, we added 'parse_name' and use it across modules. The same is done for 'types', except that the 'object' keyword should be ignored.

This change is very helpful as the input validation is done during the initialization of the PDDL objects.

The keyword validation functions have been moved to the pddl.custom_types module.